+++ examples/v2/next-pages-router/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @copilotkitnext/next-pages-router
 
+## 1.54.0-next.4
+
+### Patch Changes
+
+- @copilotkitnext/react@1.54.0-next.4
+
+## 1.54.0-next.3
+
+### Patch Changes
+
+- Updated dependencies [fa0d1cd]
+  - @copilotkitnext/react@1.54.0-next.3
+
 ## 1.53.1-next.2
 
 ### Patch Changes