+++ packages/v1/a2ui-renderer/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @copilotkit/a2ui-renderer
 
+## 1.54.0-next.4
+
+## 1.54.0-next.3
+
 ## 1.53.1-next.2
 
 ## 1.53.1-next.1