+++ packages/v1/runtime/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @copilotkit/runtime
 
+## 1.54.0-next.4
+
+### Patch Changes
+
+- @copilotkit/shared@1.54.0-next.4
+- @copilotkitnext/agent@1.54.0-next.4
+- @copilotkitnext/runtime@1.54.0-next.4
+
+## 1.54.0-next.3
+
+### Minor Changes
+
+- fa0d1cd: Add support for Standard Schema (instead of just Zod)
+
+### Patch Changes
+
+- Updated dependencies [fa0d1cd]
+  - @copilotkitnext/agent@1.54.0-next.3
+  - @copilotkit/shared@1.54.0-next.3
+  - @copilotkitnext/runtime@1.54.0-next.3
+
 ## 1.53.1-next.2
 
 ### Patch Changes