+++ packages/v2/agent/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @copilotkitnext/agent
 
+## 1.54.0-next.4
+
+### Patch Changes
+
+- @copilotkitnext/shared@1.54.0-next.4
+
+## 1.54.0-next.3
+
+### Minor Changes
+
+- fa0d1cd: Add support for Standard Schema (instead of just Zod)
+
+### Patch Changes
+
+- @copilotkitnext/shared@1.54.0-next.3
+
 ## 1.53.1-next.2
 
 ## 1.53.1-next.1