+++ packages/v2/angular/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @copilotkitnext/angular
 
+## 1.54.0-next.4
+
+### Patch Changes
+
+- @copilotkitnext/core@1.54.0-next.4
+- @copilotkitnext/shared@1.54.0-next.4
+
+## 1.54.0-next.3
+
+### Patch Changes
+
+- @copilotkitnext/core@1.54.0-next.3
+- @copilotkitnext/shared@1.54.0-next.3
+
 ## 1.53.1-next.2
 
 ### Patch Changes